+++ src/windows/AzPin.Windows.Core/Models/Entities/HiddenSubscription.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace AzPin.Windows.Models.Entities;
+
+public class HiddenSubscription
+{
+    [Key]
+    public string SubscriptionId { get; set; } = string.Empty;
+}